refactor: move finngen.py into preprocess/
