--- src/otg/datasource/finngen/__init__.py
@@ -1,3 +1,0 @@
-"""GWAS Catalog Data Source."""
-
-from __future__ import annotations
